requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
emoji==2.14.0
entrypoints==0.4
executing==2.1.0
extra-streamlit-components==0.1.71
faiss-cpu==1.9.0.post1
Faker==33.1.0
favicon==0.7.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.17.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
htbuilder==0.7.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
huggingface-hub==0.28.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.16
langchain-core==0.3.32
langchain-text-splitters==0.3.5
langsmith==0.3.2
lazy_loader==0.4
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
narwhals==1.15.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
openai==1.56.0
opencv-python-headless==4.10.0.84
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pydrive2==1.21.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
pymdown-extensions==10.12
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.0
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.14.1
sentence-transformers==3.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.40.2
streamlit-authenticator==0.4.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-google-auth==1.1.8
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
tts-wrapper @ git+https://github.com/PassionFruits-net/tts-wrapper.git@9a5a873f7ae16c79bd1c77f97417ffa196d53451
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
vulture==2.14
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0
